# implementation/platform/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: My-FinAI-Manager Platform API
  version: 0.1.0
  description: >
    External business API for My-FinAI-Manager.

    FD001 — Create Investment Portfolio adds `POST /api/portfolios`.
    FD003 — List and View Portfolio Details adds `GET /api/portfolios` (the current investor's
    saved portfolios, newest first) and `GET /api/portfolios/{portfolioId}` (one portfolio with
    its positions). Both are read-only — no sorting, filtering, searching, or pagination.
    EN004 — Establish Financial Instrument Reference Data adds `GET /api/financial-instruments`
    (the catalog search FD002 uses for controlled instrument selection).
    Editing, valuation, risk and recommendations are not part of this version.
    Operational backend/database health is exposed via the Spring Boot Actuator health mechanism at
    `/actuator/health` and is intentionally not modelled here.
servers:
  - url: http://localhost:8080
    description: Local development (started via implementation/platform/start.sh)
paths:
...
  /api/portfolios/{portfolioId}/analysis/latest:
    get:
      operationId: getLatestPortfolioAnalysis
      tags: [Portfolio]
      summary: Get the most recently requested AI Portfolio Analysis
      description: >
        Returns the latest AI-generated analysis for the portfolio — diversification, key insights,
        and structured risks — computed asynchronously from the portfolio's deterministic valuation
        (FD004) via the provider-neutral AI architecture (EN006). A portfolio that has never had an
        analysis requested returns status `NONE`. Content fields are present only when `status` is
        `COMPLETED`. Read-only.
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The latest analysis (or an explicit `NONE` placeholder).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioAnalysis'
              examples:
                none:
                  summary: Never requested
                  value: { status: NONE }
                pending:
                  summary: Requested, not yet processed
                  value: { status: PENDING, requestedAt: "2026-09-05T20:00:00Z" }
                completed:
                  summary: Completed analysis
                  value:
                    status: COMPLETED
                    requestedAt: "2026-09-05T20:00:00Z"
                    completedAt: "2026-09-05T20:00:07Z"
                    overallDiversification: { level: MODERATE, explanation: "Concentrated in Technology." }
                    keyInsights:
                      - { type: SECTOR_EXPOSURE, message: "Technology represents 76.19% of the Portfolio." }
                    risks:
                      - { type: SECTOR_CONCENTRATION, severity: HIGH, title: "Sector concentration", explanation: "..." }
                failed:
                  summary: Failed analysis
                  value: { status: FAILED, requestedAt: "2026-09-05T20:00:00Z", completedAt: "2026-09-05T20:00:03Z" }
        '400':
          description: The `{portfolioId}` path segment is not a valid UUID.
        '404':
          description: No portfolio with this id belongs to the current investor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: /problems/portfolio-not-found
                title: Portfolio not found
                status: 404
  /api/portfolios/{portfolioId}/analysis:
    post:
      operationId: requestPortfolioAnalysis
      tags: [Portfolio]
      summary: Request a new AI Portfolio Analysis (manual re-analysis)
      description: >
        Creates a new immutable analysis record and submits it for asynchronous background
        processing. Returns promptly — never waits for AI completion. Rejected with `409` when the
        latest analysis for this portfolio is already `PENDING` or `RUNNING` (duplicate-request
        prevention is enforced by the database, not only the frontend). The previous analysis, if
        any, is never modified.
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: A new analysis was accepted for background processing.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestedPortfolioAnalysis'
        '400':
          description: The `{portfolioId}` path segment is not a valid UUID.
        '404':
          description: No portfolio with this id belongs to the current investor.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: The latest analysis for this portfolio is already `PENDING` or `RUNNING`.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              example:
                type: /problems/analysis-already-in-progress
                title: Analysis already in progress
                status: 409
                detail: A Portfolio Analysis is already pending or running for this portfolio.
...
components:
  schemas:
...
    PortfolioAnalysis:
      type: object
      additionalProperties: false
      description: >
        FD005 AI-generated Portfolio analysis. No provider/model/prompt/token/cost field is ever
        exposed here — that metadata stays in the persisted row and EN006 telemetry only.
      required: [status]
      properties:
        status:
          type: string
          enum: [NONE, PENDING, RUNNING, COMPLETED, FAILED]
          description: >
            NONE — no analysis has ever been requested for this portfolio. PENDING — requested, not
            yet processing. RUNNING — processing. COMPLETED — content fields below are populated.
            FAILED — no content fields; the Investor can request a new analysis.
        requestedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        overallDiversification:
          type: object
          nullable: true
          additionalProperties: false
          required: [level, explanation]
          properties:
            level:
              type: string
              enum: [LOW, MODERATE, HIGH]
            explanation:
              type: string
        keyInsights:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: false
            required: [type, message]
            properties:
              type:
                type: string
              message:
                type: string
        risks:
          type: array
          nullable: true
          items:
            type: object
            additionalProperties: false
            required: [type, severity, title, explanation]
            properties:
              type:
                type: string
              severity:
                type: string
                enum: [HIGH, MEDIUM, LOW]
              title:
                type: string
              explanation:
                type: string
    RequestedPortfolioAnalysis:
      type: object
      additionalProperties: false
      required: [analysisId, status, requestedAt]
      properties:
        analysisId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING]
        requestedAt:
          type: string
          format: date-time
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Stable, machine-readable problem identifier (URI reference).
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...